# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lilypad API",
        "description": "API for lilypad",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/receiving": {
            "get": {
                "tags": [
                    "Receiving"
                ],
                "summary": "Get receipt items by tracking number",
                "description": "Retrieves receipt items associated with a specific tracking number (now called `num`). If no tracking number is provided, it retrieves all receipt items. Additionally, you can filter by creation date using `createdBefore` and `createdAfter`.",
                "operationId": "438312671ffb4ccf2c32705f1a13331e",
                "parameters": [
                    {
                        "name": "num",
                        "in": "query",
                        "description": "Tracking number of the receipt to retrieve items for",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "createdBefore",
                        "in": "query",
                        "description": "Retrieve receipt items created before this date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-12-31"
                        }
                    },
                    {
                        "name": "createdAfter",
                        "in": "query",
                        "description": "Retrieve receipt items created after this date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-01-01"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "num": {
                                        "description": "Tracking number of the receipt",
                                        "type": "string",
                                        "example": "TN-12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Receipt items retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Receipt items retrieved successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Receipt not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Receipt not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Receiving"
                ],
                "summary": "Receives parts from a purchase order.",
                "description": "Endpoint to receive parts from a purchase order into inventory.",
                "operationId": "d4fa36b42c27f255705eb749d4e75a5d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "PONum": {
                                        "description": "The purchase order number used to order the received parts.",
                                        "type": "string",
                                        "maxLength": 25,
                                        "example": "123"
                                    },
                                    "Fulfill": {
                                        "description": "Indicates whether the line item will be fulfilled or received only.",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "VendorPartNum": {
                                        "description": "The vendor part number.",
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "1001"
                                    },
                                    "Qty": {
                                        "description": "The quantity of parts received.",
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "Location": {
                                        "description": "The location to receive the parts into.",
                                        "type": "string",
                                        "example": "Main"
                                    },
                                    "Date": {
                                        "description": "The date the parts were received.",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-10-09"
                                    },
                                    "ShippingTrackingNumber": {
                                        "description": "The shipping tracking number.",
                                        "type": "string",
                                        "example": "1001"
                                    },
                                    "ShippingPackageCount": {
                                        "description": "The package count for the shipment.",
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "ShippingCarrier": {
                                        "description": "The carrier used for shipping.",
                                        "type": "string",
                                        "example": "Delivery"
                                    },
                                    "ShippingCarrierService": {
                                        "description": "The shipping carrier's service level.",
                                        "type": "string",
                                        "example": "Express Saver"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Receiving details successfully recorded."
                    },
                    "400": {
                        "description": "Bad request. Missing or incorrect parameters."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            }
        },
...
    },
...
}